Add a Tyxml backwards compatibility comment
It appears it is possible to use the XML features of the CGC branch
with older versions of the Tyxml library, like the one packaged with
Ubuntu 12.04, by aliasing a module name. I don't recall whether we
had noticed this during the original development: probably SIFT was
already using a later Tyxml version, and while UMN's machines ran
Ubuntu 12.04, we had to compile Tyxml from source and got a newer
version anyway.