Add -concolic-stdin and -symbolic-stdin-concrete-size options
"-concolic-stdin" is the same as what Qiuchen implemented as
"-concolic-read" as part of commit 5ac03444 on the loop summarization
branch. -symbolic-stdin-concrete-size is the non-concolic analog; I decided
on a longer name so that in the future we could also have something
like -symbolic-stdin-unlimited or other variations that affect the size
of what is read.