Add sanity check for expression simplification in concolic mode
We check that simplification preserves the value of an expression under
the concolic valuation. This is only a partial correctness check, but
it is relatively cheap. It's particularly convenient in the case where
you already suspect a problem because a concolic execution is behaving
differently for the corresponding concrete execution.