Fix some uses of == and != in OCaml that should be = and <> respectively
This is an easy mistake to make when going back and forth between
OCaml and C-family languages, but it can lead to some relatively
subtle bugs. When applied to "int"s I think it is probably harmless,
but I think for consistency it's best to always use <> unless
using pointer comparison is an important optimization.