Automatically detect the CPU architecture from an ELF file
A frequently-suggested usability feature that removes the need for the
"-arch" option in many cases. It was the request from Ian Kariniemi
that inspired this implementation.

As part of this change, FuzzBALL no longer defaults to x86 when -arch
is omitted. In the less common cases where there is no ELF file
(including for test_insn), it must now always be supplied.