Treat out-of-bounds addresses as SimulatedSegfaults
These start to occur a lot if you have address-overwrite attacks
against 64-bit programs: any address outside the low 4GB is illegal
as far as FuzzBALL is concerned. At the moment we're still not
tracking the mapping status of pages within the low 4GB, but that
could use the same exception once it's implemented.