Change the naming of -{symbolic,concolic}-{file,stdin} variables
Make them distinguished just based on the location within the sequence
of reads, without making a new base name every time a read happens.
This makes it work in a more sensible way when reads happen after the
-fuzz-start-add, including for the purposes of -trace-assigns-string.
Note though that currently the filename and file offset are not
included in the name. This means that seek()ing in the symbolic file
make still give unexpected results.