A collection of small changes to region and table handling
I noticed a variety of failures while working on an example that is
apparently slightly different from things we've tested well in the
past. In getting it to work I made a number of small changes which
should generally be pretty conservative, though this code is messy
enough that there may still be negative side-effects lurking.

- Separate the terms in -trace-sym-addr-details with semicolons,
since just spaces could be vidually lost in other whitespace.
- Disable more region cases under -no-sym-regions.
- Discard a width of 0 in the cached as well as the uncached case.
- Reduce the maximum maximum in query_maxval, to keep it positive.
- Add some more assertions to catch table-size problems sooner.
- Add another case to handle concretely-indexed table loads that can
happen even after going through the region code path.