Implement -disqualify-on-message option
To stop execution paths based on the program printing a message that
looks like an error. Only after writing this did I go back and check
that we implemented something similar on the CGC branch under the
name -stop-on-error-msgs. Besides the port from cgc_transmit to
write(2), there are some interface and behavior differences where
I'm not sure which is better.