Tweaks to -target-guidance and docs inspired by another example
Change the place at which -target-guidance is consulted so that it
applies to more choices such as index concretization, not just
conditional jumps.
Add another special value 3.0 which is like 2.0 but without the
special case for not having seen any value yet.
Document the current state of -target-guidance, and the </dev/zero
idea that is often useful in conjuction with -symbolic-stdin-concrete-size.