Add an option -auto-late-restart for a previously-disabled feature
This is a nice optimization when it works, so make it optional for
ease of experimenting with it.
Also, add a call to note_first_branch where one was previously missing,
and a related debug message. Without this fix, the feature was not
working correctly (failing with a decision tree inconsistency) when
the first symbolic branch was concretizing a symbolic address.