Detect the need for OCaml's -unsafe-string with a configure check
This option makes modern OCaml versions compatible with pre-4.02
OCaml in allowing code that mutates "string"s, which is the current
state of the FuzzBALL codebase. The configure check deals with the
situation that the flag is needed in the latest versions but was not
present in the oldest versions we'd like to maintain compatibility
with. This replaces a patch you'd had to apply manually.

Despite this change, we probably have reached a point where switching
over to a string/buffer separation would be sensible. A majority of
FuzzBALL's code treats strings as immutable. And there are probably
no longer many users who need to run FuzzBALL on systems too old
to have OCaml 4.02 (e.g., Ubuntu before 16.04). But it annoys me that
this change is not automatically backwards-compatible, and I'm not
sure what the best approach to making it backwards-compatible would be.