Reduce unneeded dependencies in vine.ml
This is a core module, so it makes things easier if it has fewer
requirements. It appears not to need ExtList anymore, and use a
module alias instead of an "open" for the few Vine_util dependencies.
Inspired by Qiuchen's experience trying to use vine.ml in ocamldebug
printers, where dependencies are a hassle.