Implement 16-bit vector x scalar SIMD shifts
As in x86 psraw, for instance.
Generalize the sse2 test program for sub-opcodes and immediates, as
used by this instruction.
Fix an incorrect #ifdef: Iop_SarN32x4 is as old as the other original
SIMD vector x scalar shift operators, I think I'd been confused
because it was on a line that was changed at the same time another op
was added.