Reduce the number of superfluous temps used in GetI/PutI translation
Switch from using the older emit_ite, which made a temp because it
used to use a multiple-statement translation, to the newer _ex_ite,
which always just makes a single expression. This makes the IR for x87
floating point instructions less gigantic looking.
It's not ideal that this code is cloned between i386 and x64.