Save and restore host FD information across iterations
This should improve FuzzBALL's behavior when the symbolic iteration
region contains opens and closes that don't match with each other.
In particular it should be less likely to run out of either host
FDs or entries in FuzzBALL's FD table, but this commit also changes
the latter to lead to a simulated system call error rather than
a crash.
Also, add code so that pipe(2) and pipe2(2) can return a simulated
EMFILE; previously the code wasn't expecting them to ever fail.
As we continue to add "snap_" backup fields to fd_info, it's looking
more desirable to refactor the snapshotting mechanism to keep a separate
copy of the whole array rather than duplicating individual fields.
Inspired by a bug report from Ian Kariniemi.