Use non-polymorphic compare in some simplification uses
The use in sorting sub-expressions is hot, and seems like it could be
related to a bug I've heard about. The uses in Frag_simplify seemed to
be among the hotter uses in a sample run I was profiling with.
Together these changes appear to be a modest performance win on that
example.