Re-simplify after re-collapsing temps to ensure canonicalization
The expression simplification we inherit from Vine also sorts
sub-expressions to make canonical representatives, but this was
being subtly broken in some cases where we trying to simplify inside
temporary variables. The reason this caused a problem is that we
were generating equivalent but un-equal (as in "a + b" vs. "b + a")
expressions on different execution paths, leading to differences in
concretization caching and decision tree inconsistencies.
Just re-doing the simplification after re-collapsing the temps
seems like an architecturally clean solution, though it does
have a measurable performance cost.