Avoid reusing novel t-variables across iterations
This adds a separate data structure to keep track of which subexpression
t variables are "enabled" on the current iteration. Even though the
numering and values of the variables are still reused across iterations,
A variable won't be used until its expression has been encountered.
The previous behavior without this restriction could lead to
inconsistent simplification when a code prefix was re-executed after
new variables were discovered, occasionally leading to a decision
tree inconsistency. Colleagues from SIFT reported the test case that
revealed this problem and participated in brainstorming about it.