Update multiple-shift simplification rules to avoid large shift amounts
We formerly had rules that added the shift amounts for repeated shifts.
which created shift amounts larger than the word size. For
zero-extending shifts these were usually later rewriten to 0 by
another rule, but this didn't happen for @>>, and it seems cleaner to
have the shift amount always be in range as an invariant. So replace
the rules with two cases which either add to a combined shift amount
in range, or rewrite to 0 right away.