Fix < 0 check in MullS64 translation
We translate 64x64->128 expanding signed multiplication by doing
a case analysis on signs and possibly negating the result of an
unsigned expanding multiply on the absolute values. But the formula
that was checking for the inputs being negative was incorrectly using
an unsigned "< 0" check, which was always false. Fix to use the
correct signed comparison instead. I discovered this bug via a weird
effect: it was causing the day of the month to be off by one when
file timestamps were converted to a string, because the compiler
implemented year/100 in a Gregorian leap-year calculation in
localtime(3) using multiplication by 0.64*2**64.